core-requirements.txt
fastapi==0.115.0
opensearch-haystack==1.0.0
uvicorn==0.30.1
redis==5.0.8
apscheduler==3.10.4
python-dotenv==1.0.1
boto3== 1.35.23
# thirdai==0.9.12
pymongo==4.9.1
groq==0.11.0
motor==3.6.0
httpx==0.27.2
sentence-transformers>=3.0.0
python-multipart==0.0.10
langchain-groq==0.2.0
langchain==0.3.3
pyjwt
